refactor: KEYS & ARMS Phase 3 アーキテクチャ再構築
- 各ステージを background/logic/renderer/index の4ファイル構成に分割
  - cave: index.ts (549行→21行) → cave-background + cave-logic + cave-renderer
  - prairie: index.ts (567行→21行) → prairie-background + prairie-logic + prairie-renderer
  - boss: index.ts (1022行→21行) → boss-background + boss-logic + boss-arena-renderer + boss-scene-renderer + boss-renderer
- engine.ts を 339行→197行に削減
  - core/input.ts: 入力システムを DI パターンで分離
  - core/game-state.ts: ゲーム状態初期化を委譲（localStorage に try-catch 追加）
  - core/render-effects.ts: 描画エフェクトを抽出
- デザインパターン導入
  - Strategy: cave-hazard-registry, prairie-enemy-registry
  - Observer: stage-event-integration (SFX トリガー、遷移ハンドラー)
- テスト追加: 6ファイル 53テスト（input, game-state, render-effects, hazard-registry, enemy-registry, event-integration）
- レビュー指摘対応: 未使用インポート削除、parseInt radix、localStorage 安全化
- tasks.md 更新: Phase 3 完了、Phase 4 に型安全性改善タスク追加、Phase 6 に DRY 改善タスク追加
